Don't try to build sqlite when cgo isn't enabled
